requirements.txt
streamlit
faiss-cpu
openai
pickle5
numpy